Fix tests compilation
